requirements.txt
Django==1.7
djangorestframework
psycopg2
